WIP: Blender variables
Add a parameterless "apply variables to path" function that does nothing
so far, and call it everywhere that `BLI_path_abs()` is also called.
This is probably not 100% correct, but sets me up to start implementing
the function in a stupid way first just to get a baseline stupid
implementation of variable substitution.