Point Cloud: Draw selection overlay
This uses a simpler point representation for selection.

There is still debate over whether we should use the same shape that is
used to draw the points themselves (and account for their radius).

If we were to do that we would need to have a minimum draw size.

The current solution is supposed to be faster than the point shape.
And we can always reiterate on this after this is merged.

---

Note: The `overlay_edit_pointcloud_vert.glsl` file shows as modified here because it was
accidently committed as part of 1584cd9aa50. It is to be fully reviewed in this patch.