Fix #128966: Update Dopesheet playhead in separate window
The playhead is redrawn as an overlay, so instead of a full region redraw using
ARegionType.draw(), at least an overlay only redraw using
ARegion.draw_overlay() needs to be triggered.

Any redrawing within a window is skipped if neither the screen, nor any of its
areas or regions are tagged for redraw. So since there are no other areas or
regions to be fully redrawn in this window, no redrawing will happen. Tagging
the screen for redraw ensures at least the overlays are redrawn, even if no
region is fully redrawn.