Fix #134491: Adjust Text Input Offset When Deleting Selection
When a text string is longer than the available space in an input it
can be at a scrolled offset, so that you can edit text that is longer
than its container is wide. When selected text is deleted we don't
update this offset. In some cases this can result in no text being
shown at at all. This PR decreases the offset by the amount of the
selected string that is currently scrolled out of view.