PyAPI: KeyMap.find_user_from_addon(...)
There was no convenient way for an add-on to show it's "user"
keymap items in the preferences, "user" meaning the key-map
item which is edited in the key-map editor, not the original
key-map item which should be created in `wm.keyconfigs.addon`.

Add a method to key-maps to lookup a user key-map item from an
add-on keymap & keymap-item.