Category: Geometry Node `Curve Spiral` primative input socket re-type.
`Curve Spiral` is one of the few nodes with `Resolution` not defined as
the number of control points in the resulting curve, but instead as a
number of points per "rotation". Since the final number of points is
defined as the rounded product of the resolution and the rotations,
non-integer values should be allowed for more finer control.

Integer `Resolution` types were changed to float, and rounding was
added to account for floating point errors.