Fix #134482: Grease Pencil interpolation leaves uninitialized values
This was partially fixed in ef844bad but left some unhandled corner cases
with uninitialized memory.

The core problem is that the `dst_curve_mask` used by the interpolation
functions can contain indices that don't actually exist in either the "from"
or "to" source drawings. Specifically when the "from" drawing has more curves,
the indices are still used but the supplemental `to_curve_indices` array
contains `-1` entries, indicating that only the "from" curve should be used.

The main copy code for positions takes this into account, but the generic
copy of attributes below that does not! It passes the `dst_curve_mask` straight
into the `array_utils` functions and that causes crashes. The original fix in ef844bad
was to clamp the size of the index mask, but the way this was done could lead to empty
index masks, leaving attribute values uninitialized.

The correct solution is to use index masks that exclude invalid entries for the
respective source curves. The new masks are non-overlapping, so the full set of
destination curves can be filled by combining these masks.