'Dummy' menu to search and de-nesting operatos
Rather than calling an operator within an operator, which leads to bad UX for undo, now using the Menu Search, but the menu that is being searched is a 'Dummy' menu that instead calls the node swap operator.

Needs to be fixed because having a 'dummy' menu is definitely not a solution, but a working version now.