Compositor: Track last update ID for cache invalidation
This is a prototype implementation of static cache invalidation through
tracking of the last update ID. Images and other resources that can be
cached now have a last_update runtime member that is initially zero and
gets incremented every time it is updated. Resources store a copy of the
last update ID when they cached their data, and invalidate the data if
it changed.

Compared to #134878, this is simpler and more robust, since update IDs
are isolated to the resource itself and not to the DEG update count.

This is just a prototype, other resources still need to be supported and
the old DrawData mechanism needs to be removed.