Fix #134954: Mask by Color doesn't work
Mistake in a0e6e16da5ed03c2a00d74154ed21b6d5e67fe24

Testing of the previous commit was inadequate, it works when a brush
tool is selected and invoked via the operator search, as the
`active_vert` property will be set by the raycast. When the Mask by
Color tool is actually active, the brush cursor, and thus the function
that performs a racyast is inactive, preventing the operator from
running.