Fix #134941: GPU: Performance regression
A performance regression was detected after 3ca5f6f62e35b18df90c62a383de5a66d41611bd.
The cause is unclear as it shader compilation is done as part of the
OpenGL driver. This PR works around the issue by moving the suppression
of compilation warnings to `CMakeLists.txt`.