Add explicit PathVariables struct, and appropriate function to build it
The BLI code shouldn't know or care about Blender internals, and yet
`BLI_path_appy_variables()` needs access to things like the current
blend file path, scene frame number, etc.

To resolve this, we define a struct in BLI that can hold variables and
their values in a dumb way, and then code outside of BLI can fill it in
and pass it to the function that does the actual path processing.

As a side-benefit, this also nicely separates the concern of determining
variables and their values from the concern of doing the path parsing
and variable substitution.