Fix #134810: Grease Pencil, keying Layer "Hide" option doesn't un-hide
Only remove a grease pencil layer from the evaluated copy when its
visibility is not animated.

Animation evaluation does not recreate the evaluated copy from scratch,
and thus once a layer is removed due to it being invisible, it never got
back.

The animatedness of the `.hide` property is determined once, when the
evaluated copy is created. This is subsequently used during evaluation
to determine whether the layer can be removed or not.

> [!Note]
> I've just created the runtime property, and accessed it directly without
> any accessor function in between. I'll leave it to the reviewers to
> determine whether this is ok.