Simplify loop logic, rename vars for clarity
offset and nth were constants, and depth was what causes alternation.  However none of that math is necesary -- it can simply be done with call order.

Along the way, 'seek_a' and 'seek_b' were renamed to what they acutally are -- lists of verts that are tagged for collapse and ignore, respectively.

Further, by checking if any verts were tagged during the first iteration pass, the second iteration pass can be avoided entirely if it will perform no work.