Fix: Geometry Nodes menu socket can leak memory.
The enum items propagation mechanism for Menu sockets makes a local copy
of item pointers to propagate between node group inputs and interface
definitions. In case these items are not used (e.g. because they are already
shared with the socket) the local copy must still be released to avoid
leaking the memory.

This only happens with more than 4 enum items, because of the internal
default buffer for `Vector`.