Fix #133010: Rigify operators for copying rig-type parameters fail
The underlying issue is was that the code was trying to directly
assign a Python dictionary to a Group property. Presumably(?) this was
supported at some point, but at least at present it no longer is and
throws an exception due to the type mismatch.

The fix is to use the Group property's `update()` method, which takes a
Python dictionary and updates any matching sub-properties with the
values in the dictionary.

Co-authored-by: Ayoub ibn Muhammad