Fix #135235: Grease Pencil: Crash when removing layers in python API
When iterating over the layers in a Grease Pencil object,
removing the layers can lead to crashes.

While this way of removing layers is not going to work, Blender
shouldn't crash. Ideally, it would report a good runtime error.

The fix adds a few checks that make sure the pointers to the
iterator element is still valid, by checking if they are contained
in the appropriate array.

This does mean that we have to do a linear search for every
iterator access leading to quadratic time complexity when
iterating over all layers.