Nodes: Make automatic viewer node positioning toggleable by user preference
When the user views an output socket
(normally Ctrl Shift + Left Click),
the Viewer node automatically repositions
itself near the clicked output.

This isn't necessarily desired
behavior for all users, so a property is added
in the User Preferences to toggle this behavior.
(It's set to be on by default.)