Fix #135163: Grease Pencil layer visibility animation doesn't work in NLA
Add a new function `bke::animdata::prop_is_animated()` that returns
whether an RNA path + array index is animated by anything. This covers
the assigned Action, NLA Action strips, and drivers.

The easiest way to implement this with the existing F-Curve-visiting
logic was to call `adt_apply_all_fcurves_cb()`. However, that function did
not allow the callback function to signal "stop looping, I found what was
I was looking for", so I extended it to do just that. I don't expect the
extra conditions to significantly slow down the other uses, as the branch
predictor will very likely optimise for the "returning true" case.