Start simpler
Rather than trying to build the variables for every call site, I'm
pulling back and just focusing on the render use case first.

I'm still leaving all the calls to `BLI_path_apply_variables()` in
place, however, because that will make it much easier to track down
where we need to compute variables for other use cases later, by setting
a break point in that function in a debugger.