UI: Introduce Playback Controls footer
> [!NOTE]
> This could be part of Blender 5.0, or as transition for 4.5 LTS
> before the Timeline editor is removed (#131102).

----

Add a new region "Playback Controls" to the following editors:
- Dope Sheet
- Graph
- NLA
- Sequencer

It's hidden by default. Available in the View menu in each editor.

Note: It could also be added to the Movie Clip editor in
DopeSheet/Graph mode, but for simplicity as initial patch I decided to
focus to the traditional editors.