Fix #135312: GPv3 paint operators disallowed during animation playback.
The paint operator crashes because the active smoothing window becomes
larger than the current drawing size when the active frame is changing
during operator execution.

Paint operators in GPv3 may cache data that is specific to the drawing
on the frame the operator started in. Changing the scene frame
invalidates much of this data, such as the smoothing window, original
and intermediate 2D positions, etc.

Making the operator robust enough to handle changes of the scene frame
is a larger task and was not anticipated during initial design. It
should just not be allowed while the operator is running.