Fix #61298: Improved floating point precision of number inputs
Fixes a long standing problem with floating point point precision
when clicking in and out of number inputs. This PR tries to only
change the string that is given back to the user as editable text,
no other precision changes in the UI should be visible.

Also fixes the precision for copy-pasting of values from one field
to another.