Fix #134989: Do Not Allow Node Deletion While File Browsing
A simplistic way of addressing this issue, but possibly harmless while
waiting for a better solution. Adds a function that returns whether we
currently have a File Browser, browsing for files, and with an operator
set to be invoked with "load" button. This is used to stop any node
deletion and return an error report.